Implemented ONSResponse class
